requirements.txt
pyaudio
sounddevice
simpleaudio
pydub
scipy
speechrecognition
google-cloud-texttospeech
Pillow
